+++ packages/smarthr-ui/src/components/Checkbox/client/ActualCheckbox.tsx
@@ -40,6 +40,7 @@ const callbackRef = (node: HTMLInputElement | null) => {
 
 export const ActualCheckbox: FC<Props> = ({ checkboxRef, checked, mixed, error, ...rest }) => {
   const mergedRef = useMergeRefs(callbackRef, checkboxRef)
+  const errorAttr = error || undefined
 
   return (
     <input
@@ -47,7 +48,8 @@ export const ActualCheckbox: FC<Props> = ({ checkboxRef, checked, mixed, error, 
       ref={mergedRef}
       type="checkbox"
       checked={checked}
-      aria-invalid={error || undefined}
+      aria-invalid={errorAttr}
+      data-smarthr-ui-input-error={errorAttr}
       data-smarthr-ui-input="true"
       // checkedはDOM属性ではなくプロパティとしてのみ反映されるため、data-checkedをMutationObserverで監視
       data-checked={checked || undefined}