+++ packages/smarthr-ui/src/components/Dialog/FocusTrap.tsx
@@ -6,10 +6,8 @@ import {
   forwardRef,
   useImperativeHandle,
   useMemo,
-  useRef,
 } from 'react'
 
-import { useMergeRefs } from '../../hooks/client/useMergeRefs'
 import { tabbable } from '../../libs/tabbable'
 
 type Props = PropsWithChildren<{
@@ -24,11 +22,9 @@ const DUMMY_FOCUS_CLASSNAME = 'smarthr-ui-Dialog-dummyFocus'
 const DUMMY_FOCUS_SELECTOR = `.${DUMMY_FOCUS_CLASSNAME}[tabIndex]`
 
 export const FocusTrap = forwardRef<FocusTrapRef, Props>(({ firstFocusTarget, children }, ref) => {
-  // TODO: innerRefを削除して、functionsのuseMemoの中にletで変数としてnodeの参照を持つことを検討
-  const innerRef = useRef<HTMLDivElement | null>(null)
-
   const functions = useMemo(() => {
-    const findDummyFocus = () => innerRef.current?.querySelector<HTMLElement>(DUMMY_FOCUS_SELECTOR)
+    let inner: HTMLDivElement | null = null
+    const findDummyFocus = () => inner?.querySelector<HTMLElement>(DUMMY_FOCUS_SELECTOR)
 
     const focus = () => {
       ;(firstFocusTarget?.current || findDummyFocus())?.focus()
@@ -36,6 +32,8 @@ export const FocusTrap = forwardRef<FocusTrapRef, Props>(({ firstFocusTarget, ch
 
     return {
       callbackRef: (node: HTMLDivElement | null) => {
+        inner = node
+
         if (!node) {
           return
         }
@@ -77,6 +75,7 @@ export const FocusTrap = forwardRef<FocusTrapRef, Props>(({ firstFocusTarget, ch
         window.addEventListener('keydown', handleKeyDown)
 
         return () => {
+          inner = null
           cancelAnimationFrame(rAFId)
           window.removeEventListener('keydown', handleKeyDown)
 
@@ -89,12 +88,10 @@ export const FocusTrap = forwardRef<FocusTrapRef, Props>(({ firstFocusTarget, ch
     }
   }, [firstFocusTarget])
 
-  const mergedRef = useMergeRefs(innerRef, functions.callbackRef)
-
   useImperativeHandle(ref, () => functions as { focus: () => void }, [functions])
 
   return (
-    <div ref={mergedRef}>
+    <div ref={functions.callbackRef}>
       {!firstFocusTarget && (
         /* eslint-disable-next-line smarthr/a11y-scroller-has-tabindex -- dummy element for focus management. */
         <div tabIndex={-1} className={DUMMY_FOCUS_CLASSNAME} />