+++ packages/smarthr-ui/src/components/Dropdown/client/DropdownMenuButton/DropdownMenuButton.tsx
@@ -163,7 +163,6 @@ export const DropdownMenuButton: FC<Props> = ({
 
       // HINT: tabとarrow keyで挙動を揃えるため、tabもhandling対象にする
       if (e.key === 'Tab') {
-        // HINT: tbのデフォルトの挙動の場合のみ、preventDefaultが必要
         e.preventDefault()
         direction = e.shiftKey ? -1 : 1
       } else if (KEY_UP_REGEX.test(e.key)) {