+++ packages/smarthr-ui/src/components/Table/client/TableReel.tsx
@@ -6,12 +6,11 @@ import {
   type PropsWithChildren,
   useCallback,
   useMemo,
-  useState,
 } from 'react'
 import { tv } from 'tailwind-variants'
 
 import { useAnimationFrame } from '../../../hooks/client/useAnimationFrame'
-import { TableScroller } from '../TableScroller'
+import { ScrollerSwitcher } from '../ScrollerSwitcher'
 import { reelShadowClassNameGenerator } from '../reelShadowStyle'
 
 const TR_SELECTOR = 'table tr'
@@ -33,15 +32,17 @@ const classNameGenerator = tv({
 })
 
 export const TableReel: FC<Props> = ({ className, children, fixedHead, ...rest }) => {
-  // TODO: stateではなくdata属性などを直接変更することで再レンダリングを引き起こさない形にしたい
-  const [showShadow, setShowShadow] = useState(false)
-
   const frame = useAnimationFrame()
 
   const callbackRef = useCallback(
     (node: HTMLElement | null) => {
       if (!node) {
         return
+      }
+
+      const wrapperNode = node.querySelector<HTMLElement>('.smarthr-ui-TableReel')
+      const setShowShadow = (visible: boolean) => {
+        wrapperNode?.setAttribute('data-smarthr-ui-shadow', visible.toString())
       }
 
       const handleScroll = () => {
@@ -139,18 +140,18 @@ export const TableReel: FC<Props> = ({ className, children, fixedHead, ...rest }
     const { wrapper, inner } = classNameGenerator()
 
     return {
-      wrapper: reelShadowClassNameGenerator({ showShadow, className: wrapper({ className }) }),
+      wrapper: reelShadowClassNameGenerator({ className: wrapper({ className }) }),
       inner: inner(),
     }
-  }, [showShadow, className])
+  }, [className])
 
   return (
-    <TableScroller ref={callbackRef} fixedHead={fixedHead}>
+    <ScrollerSwitcher forwardedRef={callbackRef} fixedHead={fixedHead}>
       <div className={classNames.wrapper}>
         <div {...rest} className={classNames.inner}>
           {children}
         </div>
       </div>
-    </TableScroller>
+    </ScrollerSwitcher>
   )
 }